msys-core: stat/lstat: drop sub-directory counting for directories
Add a patch that removes extra counting of sub-directories in [l]stat.
This fixes stat.st_nlink for directories (always 1) and improves [l]stat
performance by at least factor 2 (for empty dirs, factor 10 for dirs with
~500 entries).

Note: programs that _rely_ on st_nlink being the number of sub-dirs + 2
need to be fixed.

Signed-off-by: Karsten Blees <blees@dcon.de>