msys-core: use canonical posix path for ambiguous mount entries
When looking up a posix path for a windows path, and the windows path is
mounted to multiple posix paths, return the canonical (i.e. shortest)
posix path instead of the longest.

E.g. when starting a shell in the root directory (i.e. the parent dir of
msys.dll), the posix path should be '/', not '/usr' or '/usr/local'.

This broke in 1.0.16 ("Be more robust when dealing with ambiguous paths
and mount points").

Signed-off-by: Karsten Blees <blees@dcon.de>